Fix swap settings update
